wagsley/pyproject.toml
[tool.poetry]
name = "wagsley"
version = "0.1.0"
description = ""
authors = ["Kurtis Fields <kurtisfields@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
wagtail = "^2.15.1"
django-colorful = "^1.3"
django-el-pagination = "^3.3.0"
django-social-share = "^2.2.1"
puput = "^1.1.3"
django-allauth = "^0.46.0"
django-crispy-forms = "^1.13.0"
crispy-bootstrap5 = "^0.6"
django-avatar = "^5.0.0"
django-comments-xtd = "^2.9.4"
ariadne-django = "^0.2.0"
lxml = "^4.6.4"
channels = "^3.0.4"
loguru = "^0.5.3"
django-cors-headers = "^3.10.0"
Markdown = "^3.3.6"
html5lib = "^1.1"

[tool.poetry.dev-dependencies]
pyppeteer = "^0.2.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
